+++ src/TypeInstances/Pair.jl
@@ -8,6 +8,4 @@ using TypeClasses
 
 TypeClasses.neutral(::Type{Pair{F, S}}) where {F, S} = neutral(F) => neutral(S)
 TypeClasses.combine(p1::Pair, p2::Pair) = combine(p1.first, p2.first) => combine(p1.second, p2.second)
-
-TypeClasses.absorbing(::Type{Pair{F, S}}) where {F, S} = absorbing(F) => absorbing(S)
 TypeClasses.orelse(p1::Pair, p2::Pair) = orelse(p1.first, p2.first) => orelse(p1.second, p2.second)